refactor: better item interactors and some shop renames for item repo

+++ src/todo_rpg/application/interactors/items/delete_current_user_item.py
@@ -5,7 +5,11 @@ from todo_rpg.application.interfaces.repositories_interfaces import (
 )
 from todo_rpg.application.interfaces.transaction_interfaces import UoWProtocol
 from todo_rpg.application.interfaces.cash_interfaces import RedisRepositoryProtocol
-from todo_rpg.application.exceptions import SessionNotFoundError, ItemNotFoundError
+from todo_rpg.application.exceptions import (
+    SessionNotFoundError,
+    ItemNotFoundError,
+    AccessDeniedError,
+)
 
 
 class DeleteCurrentUserItemInteractor:
@@ -26,5 +30,7 @@ class DeleteCurrentUserItemInteractor:
         item = await self.repo.get_item_by_id(item_id)
         if item is None:
             raise ItemNotFoundError()
+        if item.user_id != user_id:
+            raise AccessDeniedError()
         item.delete()
         await self.uow.commit()
